Revise documentation for qBittorrent theming updates
Updated the documentation to reflect changes in qBittorrent 4.6.0, including migration to Qt6 and the .qbtheme file format. Clarified usage of config.json for GUI color management and custom resources.

TODO Note: update the link to the build script, the template/python script is too outdated for the recent qbittorrent v5.1.x and later.